pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "google-genai>=1.73.1",
    "homeharvest>=0.6.0",
    "httpx>=0.27.0",
    "imagekitio>=3.2.0",
    "pillow>=10.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.24",
    "sqlalchemy>=2.0.30",
    "supabase>=2.4.0",
    "uvicorn[standard]>=0.30.1",
]
